requirements.txt
# FastAPI Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Core Dependencies
requests==2.31.0
pandas==2.1.3
numpy==1.24.4
pytz==2023.3

# Database & Supabase Integration
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
supabase==2.1.0

# Authentication & JWT
python-jose[cryptography]==3.3.0

# Configuration Management
pydantic-settings==2.7.1

# Background Tasks & Caching
celery==5.3.4
redis==5.0.1

# Development & Testing
pytest==7.4.3
python-dotenv==1.0.0

# Optional: Performance monitoring
psutil==5.9.6
